Redefinition Project: added a support framework (Maven), Eliminated some (almost all) classes
