--- src/game/properties/OpData.java
@@ -1,9 +1,0 @@
-package game.properties;
-
-import java.util.ArrayList;
-import java.util.HashMap;
-
-public class OpData {
-    private ArrayList<String> require;
-    private HashMap<String, ArrayList<String>> parameters;
-}
