--- src/game/properties/Option.java
@@ -1,17 +1,0 @@
-package game.properties;
-
-public class Option {
-    private Presentation presentation;
-
-    public Option(Presentation presentation){
-        this.presentation = presentation;
-    }
-
-    public String getNameOption() {
-        return presentation.getName();
-    }
-
-    public Description getDescription() {
-        return presentation.getDescription();
-    }
-}
