--- src/game/properties/Test.java
@@ -1,8 +1,0 @@
-package game.properties;
-
-public class Test {
-
-    public static void main(String[] args) {
-        //Not testable for now
-    }
-}
