--- src/game/story/Story.java
@@ -1,9 +1,0 @@
-package game.story;
-
-import game.properties.Scenario;
-
-import java.util.ArrayList;
-
-public class Story {
-    private ArrayList<Scenario> scenarios = new ArrayList<>();
-}
