--- src/game/stuff/ListOption.java
@@ -1,17 +1,0 @@
-package game.stuff;
-
-import game.properties.Option;
-
-import java.util.ArrayList;
-
-public class ListOption {
-    private ArrayList<Option> options = new ArrayList<>();
-
-    public void add(Option option){
-        options.add(option);
-    }
-
-    public Option[] getList(){
-        return options.toArray(new Option[0]);
-    }
-}
